backend/requirements.txt
fastapi
uvicorn
python-multipart
pandas
numpy
anthropic
python-dotenv